+++ docs/markdown/progress-notes/changelog.md
@@ -203,3 +203,4 @@
 - It is now a warning to redefine a constant.
 - Fix a bug where the size of static boxed arrays was only `length` and not `allocated-length`
 - It is now possible to call a method on a forward declared type. The forward declared type must be a basic.
+- Using `->` on a plain `pointer` or `inline-array` now generates an error instead of crashing the compiler